meta_analyst/requirements.txt
anthropic>=0.96.0
gspread>=6.0.0
google-auth>=2.30.0
notion-client>=3.0.0
